requirements.txt
subprocess.run==0.0.8
Pillow==7.1.2
opencv-python==4.1.1.26
mss==5.0.0
PyAudio==0.2.11
numpy==1.18.1
pyautogui==0.9.48

